Clean up reporting. We had to inline images and css because:
We had the issue where this worked locally, but not on the server.
If we used links for stylesheet and `img src` urls, it would attempt
to fetch them but the request timed out after 10 seconds.

This might be because of concurrentrequests, and the test server only
having 1 thread at the time. In any case, this approach to inline it
should be prettty fool proof.